+++ src/types/actionState.ts
@@ -1,0 +1,4 @@
+export interface ActionState {
+  status: boolean;
+  error: string;
+}
